+++ models/enum.attribute.go
@@ -1,6 +1,6 @@
 package model
 
 type EnumAttribute struct {
-	ID    uint32 `gorm:"type:INTEGER PRIMARYKEY"`
+	ID    uint64 `gorm:"type:INTEGER PRIMARYKEY"`
 	Value string `gorm:"type:TEXT;NOT NULL"`
 }
